+++ scripts/classes/CompatibilityHelpers.js
@@ -1,5 +1,5 @@
 const CompatibilityHelpers = (function () {
-	//v14 - Property management moved to foundry.utils
+	//fv14 - Property management moved to foundry.utils
 	function hasProperty(...args) {
 		if (game.version >= 12) {
 			return foundry.utils.hasProperty(...args);
@@ -39,13 +39,22 @@ const CompatibilityHelpers = (function () {
 		return Roll.replaceFormulaData(...args);
 		
 	}
+	function weight(w, display) {
+		if (isNaN(parseFloat(w)) && dnd5e.version.localeCompare(3.2, undefined, { numeric: true, sensitivity: 'base' }) >= 0) {
+			let d = display ? display == "imperial" ? "lb" : "kg" : w.units;
+			return dnd5e.utils.convertWeight(w.value, w.units, d);
+		}
+		return w;
+		
+	}
 	return {
 		hasProperty: hasProperty,
 		setProperty: setProperty,
 		getProperty: getProperty,
 		clamped: clamped,
 		mergeObject: mergeObject,
-		replaceFormulaData: replaceFormulaData
+		replaceFormulaData: replaceFormulaData,
+		weight: weight
 	};
 })();
 export default CompatibilityHelpers;